Create list of data for Stan fitting

Create list of data for Stan fitting

Usage

createStanDataForFitting(
  win,
  loss = NULL,
  choice,
  numDecks,
  utility,
  updating,
  temperature,
  pars = NULL,
  reg = 1,
  scale = 0.01
)

Arguments

win

Vector or wins

loss

Vector of losses

choice

Integer vector of deck choices

numDecks

Integer. Number of decks

utility

Name of utility function

updating

Name of updating function

temperature

Name of temperature function

pars

A named list of parameters. See details.

reg

Positive regularization parameter. A Beta(reg,reg) prior is placed over the range of each parameter, with larger values of reg giving a tighter concentration around the midpoint of the interval. Default value of 1 gives a uniform prior.

scale

Scaling factor for observed outcomes

Value

A list of data for stan

Details

Argument "pars" is a named list with fields:

"utility": A named list of parameter values. Names must correspond to the parameters of the utility function specified by argument "utility". Each parameter must be a vector of length 2, giving lower and upper bounds.

"updating": See above.

"temperature": See above.