Simulate performance of an IGT model

Simulate IGT performance using a specified model.

Usage

simulateIGT(
  deck = "bechara",
  n,
  utility,
  updating,
  temperature,
  pars,
  scale = 0.01,
  fullOutput = FALSE
)

Arguments

deck

Either the name of an included deck or a named list with matrices entries "win" and "loss".

n

Number of subjects to simulate

updating

Name of updating function

temperature

Name of temperature function

pars

A named list of parameters. See details.

scale

How to scale outcomes before passing to the model. Typically, outcomes are divided by 100 (i.e. scale = .01).

fullOutput

Whether to output internal state variables (i.e. deck valuations)

Value

A named list containing simulated IGT data.

Details

Argument "pars" is a named list with fields:

"utility": A named list of parameter values. Names must correspond to the parameters of the utility function specified by argument "utility".

"updating": See above.

"temperature": See above.

See also