Verify provided simulation parameters

Internal function. Verify that simulation parameters are correct, and correspond to implemented models.

Usage

verifySimulationParameters(utility, updating, temperature, pars)

Arguments

utility

Name of utility function

updating

Name of updating function

temperature

Name of temperature function

pars

A named list of parameters. See details.

Value

Nothing. Generates an error message in the event of incorrect input.

Details

Argument "pars" is a named list with fields:

"utility": A named list of parameter values. Names must correspond to the parameters of the utility function specified by argument "utility".

"updating": See above.

"temperature": See above.