Import IGT deck structure

Import one of the standard IGT deck structures included in the iowa package

Usage

importDeck(deck)

Arguments

deck

The name of an included deck (see details)

Value

A named list with fields "win" and "loss".

Details

iowa includes several standard IGT deck structures, in addition to the ability for the user to construct their own. All decks are named lists with fields:

wins: A numTrial x numDeck matrix of gains

losses: A numTrial x numDeck matrix of losses

Recall that, in the classic administration of the IGT, subjects experience both a gain and a loss when selecting a card. In cases where the subject sees only a net outcome (e.g. as in the prospect utility function), the net gain is computed internal as wins - losses, and so the user can simply place the net outcome in the "wins" slot, and specify the "losses" slot to be a matrix of zeros. Included decks are identified by a character string:

bechara: The standard deck structure used by Bechara et al. (1994). Note that the authors used decks of 40 cards, while the decks provided here contains 100 cards, consisting of 2.5 "copies" of each deck.

becharavar: A variant of the Bechara deck structure in which the loss of deck C is fixed at 50, rather than varying between (25, 50, 75).

soochow: The alternative deck structure used by Chiu et al. (2008) in which the win/loss frequency is symmetric across pairs of good and bad decks.